fix mixed type in qlinera_torch_zp

+++ auto_round_extension/torch/qlinear_torch_zp.py
@@ -333,7 +333,7 @@ class QuantLinear(nn.Module):
         out = torch.matmul(x, weights)
         out = out.to(x_dtype)
         out = out.reshape(out_shape)
-        out = out + self.bias if self.bias is not None else out
+        out = (out + self.bias).to(x_dtype) if self.bias is not None else out
         return out
